css/css-view-transitions/duplicate_tag_rejects_capture.html

<!DOCTYPE html>
<html class=reftest-wait>
<title>Shared transitions: root tag is reserved</title>
<link rel="help" href="https://github.com/WICG/view-transitions">
<link rel="author" href="mailto:vmpstr@chromium.org">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
div {
  width: 100px;
  height: 100px;
  background: blue;
  contain: paint;
}
</style>

<div id=first></div>
<div id=second></div>

<script>
promise_test(async t => {
  return new Promise((resolve, reject) => {
    first.style = "view-transition-name: target";
    second.style = "view-transition-name: target";
    let transition = document.startViewTransition();

    // Ready rejected first since invoking the dom callback is an async task.
    let readyRejected = false;
    transition.ready.then(reject, () => {readyRejected = true;});

    // The domUpdate promise resolves (since there is no callback).
    let domUpdatedResolved = false;
    transition.domUpdated.then(() => {
      assert_true(readyRejected, "ready not rejected before domUpdated");
      domUpdatedResolved = true;
    }, reject);

    // Finally finish resolves.
    transition.finished.then(() => {
      assert_true(domUpdatedResolved, "domUpdated not resolved before finish");
      resolve();
    }, reject);
  });
}, "Two different elements with the same tag rejects capture");
</script>
